% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlibs.R
\name{drop_vowels}
\alias{drop_vowels}
\title{Remove vowels}
\usage{
drop_vowels(x, replacement = " \\\\_ ")
}
\arguments{
\item{x}{a string input}

\item{replacement}{a string input that will replace vowels aeiou}
}
\value{
a string object
}
\description{
Takes character string input and replaces vowels with pattern; default is underscore.
}
\examples{
drop_vowels("Hello, World!")
drop_vowels("Hello, World!", "-")
}
