% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/story_grocer.R
\name{story_grocer}
\alias{story_grocer}
\title{A story about a stocking mishap}
\usage{
story_grocer(
  firstname,
  fruit,
  grocery_item,
  verb,
  weather,
  mode_of_transport,
  judgement
)
}
\arguments{
\item{firstname}{A name for the protagonist}

\item{fruit}{A type of fruit}

\item{grocery_item}{An item found at the grocery store}

\item{verb}{A past tense verb (e.g. "took" or "smashed")}

\item{weather}{Description of the weather (e.g. "windy" or "sunny")}

\item{mode_of_transport}{A transportation device (e.g. "bus" or "bicycle")}

\item{judgement}{A description (e.g. "fantastic" or "sad")}
}
\value{
a fascinating story
}
\description{
A story about a stocking mishap
}
\examples{
story_grocer(
  firstname = "James",
  fruit = "papaya",
  grocery_item = "toilet paper",
  verb = "cut up",
  weather = "drizzly",
  mode_of_transport = "train",
  judgement = "confusing"
)
}
