% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/story_science.R
\name{story_science}
\alias{story_science}
\title{A story about a new scientific breakthrough}
\usage{
story_science(
  journal,
  phenomenon,
  organism,
  biological_process,
  judgement,
  field
)
}
\arguments{
\item{journal}{A scientific journal}

\item{phenomenon}{A known natural phenomenon (e.g. "gravity", "the speed of light")}

\item{organism}{A species of plant, animal, fungus, or bacterium}

\item{biological_process}{Process important for the biology of the organism, (e.g. "reproduction" or "cell division")}

\item{judgement}{A description of quality (e.g. "outstanding", "breakthrough", or "shocking")}

\item{field}{A scientific field}
}
\description{
A story about a new scientific breakthrough
}
\examples{
{
story_science(
  journal = "Nature",
  phenomenon = "gravity",
  organism = "E. coli",
  biological_process = "cell division",
  judgement = "breakthrough",
  field = "nutrition"
  )
}
}
