% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlibs.R
\name{drop_words}
\alias{drop_words}
\title{Drop words to quiz knowledge of a text}
\usage{
drop_words(x, frac = 0.2)
}
\arguments{
\item{x}{a string input usually a paragraph}

\item{frac}{a fraction of words to replace with empty spaces}
}
\description{
Drop words to quiz knowledge of a text
}
\examples{
drop_words(x = "The woods are lovely dark and deep,
but I have promises to keep and miles to go before I sleep.", frac = .2)
}
