% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlibs.R
\name{remove_aeiou}
\alias{remove_aeiou}
\title{Remove vowels}
\usage{
remove_aeiou(x, replacement = " \\\\_ ")
}
\arguments{
\item{x}{a string input}

\item{replacement}{a string input that will replace vowels aeiou}
}
\value{
a string object
}
\description{
Takes character string input and replaces vowels with pattern; default is underscore.
}
\examples{
remove_aeiou("Hello, World!")
remove_aeiou("Hello, World!", "-")
}
