% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlibs.R
\name{remove_aeiou}
\alias{remove_aeiou}
\title{remove aeiou}
\usage{
remove_aeiou(x)
}
\arguments{
\item{x}{a string input}
}
\value{
a string object
}
\description{
remove aeiou
}
\examples{
remove_aeiou("Hello, World!")
}
